Extract based class for parser that detect test class based ob attributes
